python-service/requirements.txt
flask
requests
py_eureka_client
flask_prometheus_metrics
flask_mongoengine
flask_jwt_extended
flask_restplus